unisgned long long for large tests

+++ RemPAR.c
@@ -15,7 +15,7 @@ To compile:
 
 To execute:
 ./kruskal edgelist.txt res.txt
-edgelist.txt should contain one edge on each line "u v w" u and v are node id (unsigned int) and w is the edge weight (double).
+edgelist.txt should contain one edge on each line "u v w" u and v are node id (unsigned long long int) and w is the edge weight (double).
 res.txt will contained the list of the edges of the resulting tree
 */
 
@@ -28,19 +28,19 @@ res.txt will contained the list of the edges of the resulting tree
 #define NLINKS 100000000 //maximum number of edges for memory allocation, will increase if needed
 
 typedef struct {
-	unsigned s;
-	unsigned t;
+	unsigned long long s;
+	unsigned long long t;
 	//double w;
 } edge;
 
 typedef struct {
-	unsigned n;//number of nodes
-	unsigned e;//number of edges
+	unsigned long long n;//number of nodes
+	unsigned long long e;//number of edges
 	edge *edges;//list of edges
 } edgelist;
 
-//Compute the maximum of three unsigned integers.
-inline unsigned int max3(unsigned int a,unsigned int b,unsigned int c){
+//Compute the maximum of three unsigned long long integers.
+inline unsigned long long int max3(unsigned long long int a,unsigned long long int b,unsigned long long int c){
 	a=(a>b) ? a : b;
 	return (a>c) ? a : c;
 }
@@ -55,7 +55,7 @@ int cmpfunc (const void * a, const void * b){
 */
 
 edgelist* readedgelist(char* input){
-	unsigned e1=NLINKS;
+	unsigned long long e1=NLINKS;
 	edgelist *g=malloc(sizeof(edgelist));
 	FILE *file;
 
@@ -63,7 +63,7 @@ edgelist* readedgelist(char* input){
 	g->e=0;
 	file=fopen(input,"r");
 	g->edges=malloc(e1*sizeof(edge));
-	while (fscanf(file,"%u %u\n", &(g->edges[g->e].s), &(g->edges[g->e].t))==2){//, &(g->edges[g->e].w))==3) {//Add one edge
+	while (fscanf(file,"%llu %llu\n", &(g->edges[g->e].s), &(g->edges[g->e].t))==2){//, &(g->edges[g->e].w))==3) {//Add one edge
 		g->n=max3(g->n,g->edges[g->e].s,g->edges[g->e].t);
 		g->e++;
 		if (g->e==e1) {
@@ -82,15 +82,15 @@ edgelist* readedgelist(char* input){
 
 // unionfind structure :
 typedef struct {
-	unsigned n;//number of objects
-	unsigned *p;//parents
+	unsigned long long n;//number of objects
+	unsigned long long *p;//parents
 } unionfind;
 
-unionfind* allocuf(unsigned n){
-	unsigned i;
+unionfind* allocuf(unsigned long long n){
+	unsigned long long i;
 	unionfind* uf=malloc(sizeof(unionfind));
 	uf->n=n;
-	uf->p=malloc(n*sizeof(unsigned));
+	uf->p=malloc(n*sizeof(unsigned long long));
 	for (i=0;i<n;i++){
 		uf->p[i]=i;
 	}
@@ -98,8 +98,8 @@ unionfind* allocuf(unsigned n){
 }
 
 //Merge the clusters of x and y returns 1 iff x and y belonged to the same cluster, 0 otherwise.
-bool Union(unsigned x, unsigned y, unionfind *uf, omp_lock_t *lock){
-	unsigned i, tmp;
+bool Union(unsigned long long x, unsigned long long y, unionfind *uf, omp_lock_t *lock){
+	unsigned long long i, tmp;
 	bool b;
 
 	while (uf->p[x] != uf->p[y]){
@@ -139,7 +139,7 @@ bool Union(unsigned x, unsigned y, unionfind *uf, omp_lock_t *lock){
 	return 1;
 }
 
-edgelist *alloctree(unsigned n){
+edgelist *alloctree(unsigned long long n){
 	edgelist *el=malloc(sizeof(edgelist));
 	el->edges=malloc((n-1)*sizeof(edge));
 	el->n=0;
@@ -148,7 +148,7 @@ edgelist *alloctree(unsigned n){
 }
 
 edgelist* kruskal(edgelist* el){
-	unsigned i,u,v;
+	unsigned long long i,u,v;
 	edgelist* elr=alloctree(el->n);
 	unionfind *uf=allocuf(el->n);
 	omp_lock_t *lock=malloc(uf->n*sizeof(omp_lock_t));
@@ -177,9 +177,9 @@ edgelist* kruskal(edgelist* el){
 void printres(edgelist* el, char* output){
 	FILE* file=fopen(output,"w");
 	//double s=0;
-	unsigned i;
+	unsigned long long i;
 	for (i=0;i<el->e;i++){
-		fprintf(file,"%u %u\n",el->edges[i].s,el->edges[i].t);//,el->edges[i].w);
+		fprintf(file,"%llu %llu\n",el->edges[i].s,el->edges[i].t);//,el->edges[i].w);
 		//s+=el->edges[i].w;
 	}
 	fclose(file);
@@ -200,8 +200,8 @@ int main(int argc,char** argv){
 	el=readedgelist(argv[2]);
 
 
-	printf("Number of nodes = %u\n",el->n);
-	printf("Number of edges = %u\n",el->e);
+	printf("Number of nodes = %llu\n",el->n);
+	printf("Number of edges = %llu\n",el->e);
 
 	t2=time(NULL);
 	printf("- Time = %ldh%ldm%lds\n",(t2-t1)/3600,((t2-t1)%3600)/60,((t2-t1)%60));
@@ -218,7 +218,7 @@ int main(int argc,char** argv){
 	printf("Printing result in file %s\n",argv[3]);
 
 	//printres(elr,argv[3]);
-	printf("Number of edges in resulting spaning tree: %u\n",elr->e);
+	printf("Number of edges in resulting spaning tree: %llu\n",elr->e);
 	//printf("Sum of the weight in a minimum spaning tree: %le\n",s);
 
 	printf("- Time = %ldh%ldm%lds\n",(t2-t1)/3600,((t2-t1)%3600)/60,((t2-t1)%60));
